(update): adiciona arquivos para projeto com esbuild
