--- programming_languages/projects_fcc/todo_app/main_test.ts
@@ -1,6 +1,0 @@
-import { assertEquals } from "@std/assert";
-import { add } from "./main.ts";
-
-Deno.test(function addTest() {
-  assertEquals(add(2, 3), 5);
-});
