+++ programming_languages/projects_fcc/todo_app/public/main.js
@@ -1,0 +1,3 @@
+// main.ts
+var h1 = document.querySelector("h1");
+h1.innerHTML += " world!";
